requirements.txt
tensorflow==2.13.0
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.13.0
scikit-learn==1.3.0
tqdm==4.66.1
Pillow==10.0.0
mediapipe==0.10.3

# API-based detection dependencies
requests==2.31.0
boto3==1.29.7  # For AWS Rekognition
azure-cognitiveservices-vision-computervision==0.9.0  # For Azure Vision
google-cloud-vision==3.4.4  # For Google Vision API